# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart360 API - multi-tenant guest information PWA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/tenants/{slug}:
    get:
      operationId: getPublicTenant
      tags: [public]
      summary: Full published content tree for a tenant (guest view)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: lang
          in: query
          required: false
          schema: { type: string }
        - name: preview
          in: query
          required: false
          schema: { type: boolean }
      responses:
        "200":
          description: Tenant content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantContent"
        "404":
          description: Not found or unpublished
  /public/tenants/{slug}/search:
    get:
      operationId: searchPublicTenant
      tags: [public]
      summary: Full-text search across a tenant's visible content
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: lang
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"
...
  /admin/tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      responses:
        "200":
          description: Tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /admin/tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Tenant with full content tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tenant content tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantContent"
        "404":
          description: Not found
    patch:
      operationId: updateTenant
      tags: [tenants]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /admin/tenants/{id}/duplicate:
    post:
      operationId: duplicateTenant
      tags: [tenants]
      summary: Duplicate a tenant with its full section/category tree
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantDuplicateInput"
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
...
components:
  schemas:
...
    Tenant:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        customDomain: { type: ["string", "null"] }
        name: { type: string }
        subtitle: { type: ["string", "null"] }
        rating: { type: ["string", "null"] }
        reviewsCount: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        heroUrl: { type: ["string", "null"] }
        tourUrl: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        whatsapp: { type: ["string", "null"] }
        viber: { type: ["string", "null"] }
        instagram: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        mapQuery: { type: ["string", "null"] }
        wifiSsid: { type: ["string", "null"] }
        wifiPass: { type: ["string", "null"] }
        theme: { type: string }
        languages:
          type: array
          items: { type: string }
        isTemplate: { type: boolean }
        isPublished: { type: boolean }
        updatedAt: { type: string }
      required: [id, slug, name, theme, languages, isTemplate, isPublished, updatedAt]
    TenantInput:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
        subtitle: { type: string }
        fromTemplate: { type: boolean }
      required: [slug, name]
    TenantDuplicateInput:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
        copyContent:
          type: boolean
          description: Copy item contents too; false copies only the section/category structure
      required: [slug, name]
    TenantUpdate:
      type: object
      properties:
        slug: { type: string }
        customDomain: { type: ["string", "null"] }
        name: { type: string }
        subtitle: { type: ["string", "null"] }
        rating: { type: ["string", "null"] }
        reviewsCount: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        heroUrl: { type: ["string", "null"] }
        tourUrl: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        whatsapp: { type: ["string", "null"] }
        viber: { type: ["string", "null"] }
        instagram: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        mapQuery: { type: ["string", "null"] }
        wifiSsid: { type: ["string", "null"] }
        wifiPass: { type: ["string", "null"] }
        theme: { type: string }
        languages:
          type: array
          items: { type: string }
        isTemplate: { type: boolean }
        isPublished: { type: boolean }
    Section:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        key: { type: string }
        title: { type: string }
        subtitle: { type: ["string", "null"] }
        icon: { type: string }
        position: { type: number }
        isVisible: { type: boolean }
      required: [id, tenantId, key, title, icon, position, isVisible]
...
    Category:
      type: object
      properties:
        id: { type: string }
        sectionId: { type: string }
        label: { type: string }
        icon: { type: string }
        layout: { type: string }
        position: { type: number }
        isVisible: { type: boolean }
      required: [id, sectionId, label, icon, layout, position, isVisible]
...
    Item:
      type: object
      properties:
        id: { type: string }
        categoryId: { type: string }
        title: { type: ["string", "null"] }
        body: { type: ["string", "null"] }
        price: { type: ["string", "null"] }
        priceUnit: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        mapQuery: { type: ["string", "null"] }
        difficulty: { type: ["string", "null"] }
        duration: { type: ["string", "null"] }
        distance: { type: ["string", "null"] }
        open24: { type: boolean }
        hoursJson:
          type: ["string", "null"]
          description: JSON array of 7 entries Mon-Sun, each [openMin, closeMin] in minutes or null when closed; close may pass midnight
        noteType: { type: ["string", "null"] }
        noteText: { type: ["string", "null"] }
        bullets:
          type: array
          items: { type: string }
        position: { type: number }
        isVisible: { type: boolean }
        media:
          type: array
          items:
            $ref: "#/components/schemas/MediaEntry"
      required: [id, categoryId, open24, bullets, position, isVisible, media]
...
    MediaEntry:
      type: object
      properties:
        id: { type: string }
        itemId: { type: ["string", "null"] }
        tenantId: { type: ["string", "null"] }
        url: { type: string }
        alt: { type: ["string", "null"] }
        position: { type: number }
      required: [id, url, position]
...
    CategoryContent:
      allOf:
        - $ref: "#/components/schemas/Category"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Item"
          required: [items]
    SectionContent:
      allOf:
        - $ref: "#/components/schemas/Section"
        - type: object
          properties:
            categories:
              type: array
              items:
                $ref: "#/components/schemas/CategoryContent"
          required: [categories]
    TenantContent:
      allOf:
        - $ref: "#/components/schemas/Tenant"
        - type: object
          properties:
            sections:
              type: array
              items:
                $ref: "#/components/schemas/SectionContent"
          required: [sections]
    SearchResult:
      type: object
      properties:
        itemId: { type: string }
        categoryId: { type: string }
        sectionId: { type: string }
        title: { type: string }
        snippet: { type: string }
        categoryLabel: { type: string }
        sectionTitle: { type: string }
      required: [itemId, categoryId, sectionId, title, snippet, categoryLabel, sectionTitle]
...