# lib/api-zod/src/generated/types/sectionContent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart360 API - multi-tenant guest information PWA
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryContent } from './categoryContent';
import type { Section } from './section';

export type SectionContent = Section & {
  categories: CategoryContent[];
};