# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart360 API - multi-tenant guest information PWA
 * OpenAPI spec version: 0.1.0
 */

export interface Tenant {
  id: string;
  slug: string;
  /** @nullable */
  customDomain?: string | null;
  name: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  rating?: string | null;
  /** @nullable */
  reviewsCount?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  heroUrl?: string | null;
  /** @nullable */
  tourUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  viber?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  mapQuery?: string | null;
  /** @nullable */
  wifiSsid?: string | null;
  /** @nullable */
  wifiPass?: string | null;
  theme: string;
  languages: string[];
  isTemplate: boolean;
  isPublished: boolean;
  updatedAt: string;
}