# lib/api-zod/src/generated/types/tenantInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart360 API - multi-tenant guest information PWA
 * OpenAPI spec version: 0.1.0
 */

export interface TenantInput {
  slug: string;
  name: string;
  subtitle?: string;
  fromTemplate?: boolean;
}